- Updated README to describe some of the features - Renamed the library from Nucleus.Gaming to NucleusDotNet - Namespace revisions and renames to maintain consistency - Removed all mentions of Nucleus.Gaming - Cleaned entire project to remove unused usings

+++ Nucleus.Gaming/Nucleus.Gaming/IO/LZMA/Compress/LZ/LzInWindow.cs
@@ -4,7 +4,7 @@
 using System;
 
 namespace Nucleus.IO.LZMA {
-        public class InWindow {
+    public class InWindow {
         public Byte[] _bufferBase = null; // pointer to buffer with data
         System.IO.Stream _stream;
         UInt32 _posLimit; // offset (from _buffer) of first byte when new block reading must be done
@@ -21,12 +21,12 @@ namespace Nucleus.IO.LZMA {
         public UInt32 _streamPos; // offset (from _buffer) of first not read byte from Stream
 
         public void MoveBlock() {
-            UInt32 offset = (UInt32)(_bufferOffset) + _pos - _keepSizeBefore;
+            UInt32 offset = _bufferOffset + _pos - _keepSizeBefore;
             // we need one additional byte, since MovePos moves on 1 byte.
             if (offset > 0)
                 offset--;
 
-            UInt32 numBytes = (UInt32)(_bufferOffset) + _streamPos - offset;
+            UInt32 numBytes = _bufferOffset + _streamPos - offset;
 
             // check negative offset ????
             for (UInt32 i = 0; i < numBytes; i++)
@@ -46,7 +46,7 @@ namespace Nucleus.IO.LZMA {
                     _posLimit = _streamPos;
                     UInt32 pointerToPostion = _bufferOffset + _posLimit;
                     if (pointerToPostion > _pointerToLastSafePosition)
-                        _posLimit = (UInt32)(_pointerToLastSafePosition - _bufferOffset);
+                        _posLimit = _pointerToLastSafePosition - _bufferOffset;
 
                     _streamEndWasReached = true;
                     return;
